+++ src/__tests__/utils.spec.js
@@ -538,6 +538,8 @@ describe('utils', () => {
       // impossible according to the MSDN docs). This case was tested in the original extension so I
       // keep it here just in case
       ['tab without container and container not specified in pattern (edge case)', undefined, '', true],
+      ['tab without container and container specified in pattern', undefined, '<CONTAINER>', false],
+      ['tab without container and container specified in pattern', '', '<CONTAINER>', false],
       ['tab without container and container not specified in pattern', '', '', true],
       ['tab container matches container specified in pattern', 'CONTAINER', '<CONTAINER>', true],
       ['tab without container and rule \'No container\' specified in pattern', '', '<>', true],